refactor: rename packge RuleImpl -> rule 로 변경
