+++ src/main/java/chess/dao/CommandDao.java
@@ -11,6 +11,7 @@ import java.util.List;
 public class CommandDao {
     private static final int DEFAULT_START_ROUND = 1;
 
+
     private JdbcTemplate jdbcTemplate;
 
     private CommandDao(final DataSource dataSource) {
