+++ src/main/java/chess/domain/BoardGenerator.java
@@ -1,6 +1,7 @@
 package chess.domain;
 
 import chess.domain.RuleImpl.*;
+import chess.domain.rule.*;
 
 import java.util.ArrayList;
 import java.util.List;
