+++ src/test/java/chess/domain/PieceSymbolTest.java
@@ -1,6 +1,6 @@
 package chess.domain;
 
-import chess.domain.RuleImpl.Rook;
+import chess.domain.rule.Rook;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
