+++ src/test/java/chess/domain/RuleImpl/BishopTest.java
@@ -2,6 +2,7 @@ package chess.domain.RuleImpl;
 
 import chess.domain.Position;
 import chess.domain.Rule;
+import chess.domain.rule.Bishop;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
