+++ src/test/java/chess/domain/RuleImpl/KnightTest.java
@@ -2,6 +2,7 @@ package chess.domain.RuleImpl;
 
 import chess.domain.Position;
 import chess.domain.Rule;
+import chess.domain.rule.Knight;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
